flytt funksjoner for å oppdatere og slette endret utbetaling til context
